Main... ajout article done !

+++ srcEclipse/Acquisition/AppareilPhoto.java
@@ -7,7 +7,7 @@ package Acquisition;
 public class AppareilPhoto extends DispositifAcquisition {
 
 	public AppareilPhoto(String nom, String reference, String marque,
-			float prixLocationParJour, int stock, String typeObjectif,
+			double prixLocationParJour, int stock, String typeObjectif,
 			int nbPixels, String resolution) {
 		super(nom, reference, marque, prixLocationParJour, stock, typeObjectif,
 				nbPixels, resolution);
