+++ srcEclipse/Acquisition/Camera.java
@@ -6,7 +6,7 @@ package Acquisition;
 public class Camera extends DispositifAcquisition {
 
 	public Camera(String nom, String reference, String marque,
-			float prixLocationParJour, int stock, String typeObjectif,
+			double prixLocationParJour, int stock, String typeObjectif,
 			int nbPixels, String resolution) {
 		
 		super(nom, reference, marque, prixLocationParJour, stock, typeObjectif,
