+++ srcEclipse/Lumiere/PanneauLED.java
@@ -9,7 +9,7 @@ public class PanneauLED extends AccessoireLumiere{
 	private int nombreLED;
 	
 	public PanneauLED(String nom, String reference, String modele,
-			float prixLocationParJour, int nbStock, int nbLed) {
+			double prixLocationParJour, int nbStock, int nbLed) {
 		
 		super(nom, reference, modele, prixLocationParJour, nbStock);
 		this.nombreLED = nbLed;
