+++ srcEclipse/Lumiere/Reflecteur.java
@@ -5,11 +5,11 @@
 package Lumiere;
 
 public class Reflecteur extends AccessoireLumiere{
-	private float taille;
+	private double taille;
 	
 
 	public Reflecteur(String nom, String reference, String modele,
-			float prixLocationParJour, int nbStock, float taille) {
+			double prixLocationParJour, int nbStock, double taille) {
 		
 		super(nom, reference, modele, prixLocationParJour, nbStock);
 		this.taille = taille;
