+++ srcEclipse/Stabilisation/MaterielTournageStabilisation.java
@@ -7,7 +7,7 @@ package Stabilisation;
 public abstract class MaterielTournageStabilisation extends main.Article {
 
 	public MaterielTournageStabilisation(String nom, String reference,
-			String modele, float prixLocationParJour, int nbStock) {
+			String modele, double prixLocationParJour, int nbStock) {
 		
 		super(nom, reference, modele, prixLocationParJour, nbStock);
 	}
