+++ srcEclipse/Stabilisation/Trepied.java
@@ -6,10 +6,10 @@ package Stabilisation;
 
 public class Trepied extends MaterielTournageStabilisation {
 
-	private float hauteur;
+	private double hauteur;
 	
 	public Trepied(String nom, String reference, String modele,
-			float prixLocationParJour, int nbStock, int hauteur) {
+			double prixLocationParJour, int nbStock, int hauteur) {
 	
 		super(nom, reference, modele, prixLocationParJour, nbStock);
 		this.hauteur = hauteur;
