+++ srcEclipse/Stabilisation/Ventouse.java
@@ -6,10 +6,10 @@ package Stabilisation;
 
 public class Ventouse extends MaterielTournageStabilisation{
 
-	private float diametre;
+	private double diametre;
 	
 	public Ventouse(String nom, String reference, String modele,
-			float prixLocationParJour, int nbStock, float diametre) {
+			double prixLocationParJour, int nbStock, double diametre) {
 		
 		super(nom, reference, modele, prixLocationParJour, nbStock);
 		this.diametre = diametre;
